♻️ [#112] Refactor: Header에서 NotificationPopover를 컴포넌트로 분리
